Violently ripped out file/url association settings from general settings to their own tab. Looks like it works as before.
